Stores coordinates in db and uses them to find furthest match, instead of using google geocode every time
